Fix entities relations and document links

+++ Spring/src/main/java/com/example/documentsystem/services/impl/DocumentLinkServiceImpl.java
@@ -5,6 +5,7 @@ import com.example.documentsystem.entities.DocumentLinkEntity;
 import com.example.documentsystem.exceptions.UnauthorizedException;
 import com.example.documentsystem.models.DocumentLink;
 import com.example.documentsystem.services.DocumentLinkService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -17,6 +18,7 @@ import java.util.UUID;
 
 
 @Service
+@Slf4j
 @Transactional
 public class DocumentLinkServiceImpl implements DocumentLinkService {
     private final DocumentLinkRepository documentLinkRepository;
@@ -30,9 +32,11 @@ public class DocumentLinkServiceImpl implements DocumentLinkService {
 
     @Override
     public Long getDocumentIdFromLink(String token, String password) {
-        String encryptedPassword = passwordEncoder.encode(password);
-        DocumentLinkEntity documentLinkEntity = documentLinkRepository.findByTokenAndPassword(token, encryptedPassword)
+        DocumentLinkEntity documentLinkEntity = documentLinkRepository.findByToken(token)
                 .orElseThrow(() -> new UnauthorizedException("Wrong password for document link"));
+
+        if (!passwordEncoder.matches(password, documentLinkEntity.getPassword()))
+            throw new UnauthorizedException("Wrong password for document link");
 
         if (documentLinkEntity.getValidUntil() != null && documentLinkEntity.getValidUntil().isBefore(LocalDateTime.now())) {
             documentLinkRepository.delete(documentLinkEntity);
@@ -44,10 +48,12 @@ public class DocumentLinkServiceImpl implements DocumentLinkService {
 
     @Override
     public DocumentLinkEntity create(DocumentLink link) {
+        String encryptedPassword = passwordEncoder.encode(link.getPassword());
+
         DocumentLinkEntity documentLinkEntity = new DocumentLinkEntity(
                 UUID.randomUUID().toString(),
                 link.getDocumentId(),
-                link.getPassword(),
+                encryptedPassword,
                 link.getValidUntil());
 
         return documentLinkRepository.save(documentLinkEntity);
